+++ packages/nlp/lib/input-settings.js
@@ -8,6 +8,15 @@ const InputType = {
 } 
 
 /**
+ * @enum {string}
+ */
+const DialogState = {
+  CONFIRM_INTENT: 'CONFIRM_INTENT',
+  CONFIRM_SLOT: 'CONFIRM_SLOT',
+  ELICIT_INTENT: 'ELICIT_INTENT',
+  ELICIT_SLOT: 'ELICIT_SLOT'
+}
+/**
  *
  */
 class InputSettings {
@@ -17,12 +26,24 @@ class InputSettings {
   constructor(type) {
     this.type = type
 
+    /** @type {DialogState | undefined} */
+    this.dialogState = undefined
+
+    /** @type {string | undefined} */
+    this.intentToConfirm = undefined
+    
     /** @type {number | undefined} */
     this.maximumDigits = undefined
 
     /** @type {number | undefined} */
     this.minimumDigits = undefined
 
+    /** @type {string | undefined} */
+    this.slotToConfirm = undefined
+    
+    /** @type {string | undefined} */
+    this.slotToElicit = undefined
+    
     /** @type {number | undefined} */
     this.timeout = undefined
   }
@@ -55,6 +76,7 @@ class InputSettings {
   }
 }
 
+InputSettings.DialogState = DialogState
 InputSettings.InputType = InputType
 
 module.exports = InputSettings